requirements.txt
torch>=2.1.0
torchvision>=0.16.0
groundingdino-py
git+https://github.com/facebookresearch/sam2.git@2b90b9f5ceec907a1c18123530e92e794ad901a4
huggingface-hub>=0.20.0
opencv-python>=4.7.0
numpy>=1.24.0
Pillow>=9.5.0
transformers>=4.36.0
fastapi>=0.110.0
uvicorn>=0.22.0
python-multipart>=0.0.6
onnx 
onnxslim 
onnxruntime-gpu
ultralytics-thop>=2.0.0